requirements.txt
numpy==1.14.1
Pillow==6.0.0
protobuf==3.7.1
scikit-learn==0.21.2
scipy==1.0.0
six==1.11.0
sklearn==0.0
torch==1.1.0
torchfile==0.1.0
torchvision==0.3.0

certifi==2018.8.24
cycler==0.10.0
kiwisolver==1.1.0
matplotlib==3.0.3
pyparsing==2.4.0
python-dateutil==2.8.0

